requirements.txt
accelerate==0.28.0
chromadb==0.4.24
google-api-python-client==2.125.0
google-generativeai==0.5.0
langchain==0.1.15
langchain-community==0.0.32
langchain-core==0.1.41
langchain-openai==0.0.8
openai==1.14.0
polygon-api-client==1.13.5
pypdf==4.1.0
pysqlite3-binary
selenium==4.19.0
streamlit==1.33.0
webdriver-manager==4.0.1